requirements.txt
Django
easy_thumbnails
django-relationships
django-notification
django-taggit
PIL
django-piston
#### Tastpie Required
mimeparse>=0.1.3
python-dateutil==1.5
lxml
PyYAML
python-digest
biplist
django-tastypie